+++ R/pd_raw.R
@@ -54,11 +54,10 @@ pd_raw <- function(object, v, X, grid, pred_fun = stats::predict,
 #'   predictions. Otherwise, a list with two elements: `pred` (predictions)
 #'   and `grid_pred` (the corresponding grid values in the same mode as the input, 
 #'   but replicated over `X`).
-#' @param ohe Should factor output be one-hot encoded? Default is `FALSE`.
 #' @returns 
 #'   Either a vector/matrix of predictions or a list with predictions and grid.
 ice_raw <- function(object, v, X, grid, pred_fun = stats::predict, 
-                    pred_only = TRUE, ohe = FALSE, ...) {
+                    pred_only = TRUE, ...) {
   D1 <- length(v) == 1L
   n <- nrow(X)
   n_grid <- NROW(grid)
@@ -79,7 +78,7 @@ ice_raw <- function(object, v, X, grid, pred_fun = stats::predict,
   }
   
   # Calculate matrix/vector of predictions
-  pred <- prepare_pred(pred_fun(object, X_pred, ...), ohe = ohe)
+  pred <- prepare_pred(pred_fun(object, X_pred, ...))
   
   if (pred_only) {
     return(pred)