+++ R/utils_input.R
@@ -1,21 +1,23 @@
 #' Prepares Predictions
 #' 
-#' Converts predictions to vector, matrix or factor.
+#' Converts predictions to vector or matrix.
 #' 
 #' @noRd
 #' @keywords internal
 #' 
 #' @param x Object representing model predictions.
-#' @param ohe If `x` is a factor: should it be one-hot encoded? Default is `FALSE`.
-#' @returns Like `x`, but converted to matrix, vector, or factor.
-prepare_pred <- function(x, ohe = FALSE) {
+#' @returns Like `x`, but converted to matrix or vector.
+prepare_pred <- function(x) {
   if (is.data.frame(x) && ncol(x) == 1L) {
     x <- x[[1L]]
   }
-  if (ohe && is.factor(x)) {
-    return(fdummy(x))
+  if (!is.vector(x) && !is.matrix(x)) {
+    x <- as.matrix(x)
   }
-  if (is.vector(x) || is.matrix(x) || is.factor(x)) x else as.matrix(x)
+  if (!is.numeric(x) && !is.logical(x)) {
+    stop("Predictions must be numeric!")
+  }
+  return(x)
 }
 
 #' Prepares Group BY Variable
@@ -93,13 +95,13 @@ prepare_w <- function(w, X) {
 #' 
 #' @noRd
 #' @keywords internal
-#' @param y Vector/matrix-like of the same length as `X`, or column names in `X`.
+#' @param y Vector, factor, or matrix-like of the same length as `X`,
+#'   or column names in `X`.
 #' @param X Matrix-like.
-#' @param ohe If y is a factor: should it be one-hot encoded? Default is `FALSE`.
 #' 
 #' @returns A list with "y" (vector, matrix, or factor) and "y_names" (if `y`
 #'   was passed as column names).
-prepare_y <- function(y, X, ohe = FALSE) {
+prepare_y <- function(y, X) {
   if (NROW(y) < nrow(X) && all(y %in% colnames(X))) {
     y_names <- y
     if (is.data.frame(X) && length(y) == 1L) {
@@ -108,9 +110,17 @@ prepare_y <- function(y, X, ohe = FALSE) {
       y <- X[, y]
     }
   } else {
+    if (is.data.frame(y) && ncol(y) == 1L) {
+      y <- y[[1L]]
+    }
     stopifnot(NROW(y) == nrow(X))
     y_names <- NULL
   }
-  list(y = prepare_pred(y, ohe = ohe), y_names = y_names)
+  if (!is.vector(y) && !is.matrix(y) && !is.factor(y)) {
+    y <- as.matrix(y)
+  }
+  if (!is.numeric(y) && !is.logical(y) && !is.factor(y)) {
+    stop("Response must be numeric (or factor.)")
+  }
+  list(y = y, y_names = y_names)
 }
-