+++ NEWS.md
@@ -1,3 +1,14 @@
+# hstats 1.2.0
+
+## Major changes
+
+- Factor-valued predictions are no longer possible.
+- Consequently, also removed "classification_error" loss.
+
+## Minor changes
+
+- Code simplifications.
+
 # hstats 1.1.2
 
 ## ICE plots